remove CatalogMetadataRepository interface and rename implementation
